add some security

+++ app/Http/Controllers/TodosController.php
@@ -52,6 +52,10 @@ class TodosController extends Controller
      */
     public function update(Request $request, Todo $todo)
     {
+        if($todo->user_id != auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $data = $request->validate([
             'title' => 'required|string',
             'done' => 'required|boolean',
@@ -70,6 +74,10 @@ class TodosController extends Controller
      */
     public function destroy(Todo $todo)
     {
+        if($todo->user_id != auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+        
         $todo->delete();
         return response('Deleted todo item', 200);
     }
